+++ src/charactersData/expandedMaterialInfo.ts
@@ -5,7 +5,8 @@ import {CROWN_OF_INSIGHT, MORA} from './materialNames';
 export enum Region {
   MONDSTADT = 'monstadt',
   LIYUE = 'liyue',
-  INAZUMA = 'inazuma'
+  INAZUMA = 'inazuma',
+  ENKANOMIA = 'enkanomia'
 }
 
 const Mora = {
@@ -224,12 +225,17 @@ const MaguuKenki = {
   map_link: 'pin=-4000|-2340|Maguu Kenki'
 }
 
-
 const GoldenWolflord = {
   name: 'golden_wolflord',
   region: Region.INAZUMA,
   original_resin: 40,
   map_link: 'pin=-6720|-2660|Golden Wolflord'
+}
+
+const BathysmalVishapHerd = {
+  name: 'bathysmal_vishap_herd',
+  region: Region.ENKANOMIA,
+  original_resin: 40,
 }
 
 const crownOfInsight = {
@@ -315,6 +321,7 @@ export const ExpandedMaterialInfo: IDungeon = {
   brilliant_diamond_chunk: {...brilliantDiamondChunk},
   brilliant_diamond_gemstone: {...brilliantDiamondGemstone},
   riftborn_regalia: {...GoldenWolflord},
+  dragonheirs_false_fin: {...BathysmalVishapHerd},
 }
 
 // order is important