+++ CHANGELOG.md
@@ -30,6 +30,8 @@
   * Added characters:
     * Yun Jin
     * Shenhe
+  * Added monster info:
+    * Bathysmal Vishap Herd
 
 #### 1.1.4 December-2021
 * Changes